ConfigObj new line after section, as per issue 8.
Private repo of Stefan Parviainen has a fix for issue 8 of the public
repo to add a newline after each section to increase the readability of
the ini files. I'll just change this single issue by hand instead of
pulling his experimental branch to not pull in weird other issues.