Move template handling to tracker adding methods
Moved handling of template specifics to the adding functions of tracker.
There is no link the the Qt GUI yet, so tabs are not created for new
features, OOIs and ROIs. Adding new those other than via loading of
templates is non-functional, too.

Also further renaming of "led" to "feature" to reflect future
generalization of things to track.